+++ lib/has_configuration.rb
@@ -23,7 +23,7 @@ module HasConfiguration #:nodoc:
     #   # => loads setting without environment processing from the
     #   #    file #{self.class.name.downcase}.yml
     #
-    #   has_configuration :file => Rails.root.join('config', 'example.yml'), :env => 'staging'
+    #   has_configuration file: Rails.root.join('config', 'example.yml'), env: 'staging'
     #   # => loads settings for staging environment from RAILS_ROOT/config/example.yml file
     #
     # ==== YAML File Example